+++ ecies/utils/elliptic.py
@@ -1,9 +1,8 @@
 from coincurve import PrivateKey, PublicKey
 from coincurve.utils import get_valid_secret
 
-from ..config import ECIES_CONFIG, Config
+from .hash import derive_key
 from .hex import decode_hex
-from .symmetric import derive_key
 
 
 def generate_key() -> PrivateKey:
@@ -22,12 +21,14 @@ def generate_key() -> PrivateKey:
 
 def generate_eth_key():
     """
+    Note: `eth-keys` needs to be installed in advance.
+
     Generate a random `eth_keys.keys.PrivateKey`
 
     Returns
     -------
     eth_keys.keys.PrivateKey
-        An ethereum key
+        An ethereum flavored secp256k1 key
 
     """
     from eth_keys import keys
@@ -55,12 +56,6 @@ def hex2pk(pk_hex: str) -> PublicKey:
     return PublicKey(compat_eth_public_key(decode_hex(pk_hex)))
 
 
-def compat_eth_public_key(data: bytes):
-    if len(data) == 64:  # eth public key format
-        data = b"\x04" + data
-    return data
-
-
 def hex2sk(sk_hex: str) -> PrivateKey:
     """
     Convert ethereum hex to `coincurve.PrivateKey`
@@ -79,11 +74,30 @@ def hex2sk(sk_hex: str) -> PrivateKey:
     return PrivateKey(decode_hex(sk_hex))
 
 
+def compat_eth_public_key(data: bytes):
+    """
+    Convert ethereum public key to uncompressed public key
+
+    Parameters
+    ----------
+    data: bytes
+        Ethereum public key (64 bytes)
+
+    Returns
+    -------
+    bytes
+        Uncompressed public key (65 bytes)
+
+    """
+    if len(data) == 64:
+        data = b"\x04" + data
+    return data
+
+
 # private below
 def encapsulate(
-    private_key: PrivateKey, peer_public_key: PublicKey, config: Config = ECIES_CONFIG
+    private_key: PrivateKey, peer_public_key: PublicKey, is_compressed: bool = False
 ) -> bytes:
-    is_compressed = config.is_hkdf_key_compressed
     shared_point = peer_public_key.multiply(private_key.secret)
     master = private_key.public_key.format(is_compressed) + shared_point.format(
         is_compressed
@@ -92,9 +106,8 @@ def encapsulate(
 
 
 def decapsulate(
-    public_key: PublicKey, peer_private_key: PrivateKey, config: Config = ECIES_CONFIG
+    public_key: PublicKey, peer_private_key: PrivateKey, is_compressed: bool = False
 ) -> bytes:
-    is_compressed = config.is_hkdf_key_compressed
     shared_point = public_key.multiply(peer_private_key.secret)
     master = public_key.format(is_compressed) + shared_point.format(is_compressed)
     return derive_key(master)